requirements.txt
together==0.2.6
pdf2image==1.16.3
rich==13.7.0
matplotlib==3.8.2
python-dotenv==1.0.0 